Toward a fix for ol.tilegrid
