+++ examples/osm-vector-tiles.js
@@ -7,6 +7,7 @@ goog.require('ol.source.VectorTile');
 goog.require('ol.style.Fill');
 goog.require('ol.style.Stroke');
 goog.require('ol.style.Style');
+goog.require('ol.tilegrid');
 
 
 var format = new ol.format.TopoJSON();
