+++ examples/vector-osm.js
@@ -12,6 +12,7 @@ goog.require('ol.style.Circle');
 goog.require('ol.style.Fill');
 goog.require('ol.style.Stroke');
 goog.require('ol.style.Style');
+goog.require('ol.tilegrid');
 
 var styles = {
   'amenity': {
