+++ src/ol/source/tile.js
@@ -1,14 +1,15 @@
 goog.provide('ol.source.Tile');
 goog.provide('ol.source.TileEvent');
 
-goog.require('ol.events.Event');
 goog.require('ol');
 goog.require('ol.TileCache');
 goog.require('ol.TileState');
+goog.require('ol.events.Event');
 goog.require('ol.proj');
 goog.require('ol.size');
 goog.require('ol.source.Source');
 goog.require('ol.tilecoord');
+goog.require('ol.tilegrid');
 
 
 /**
