+++ src/ol/source/tilejson.js
@@ -14,6 +14,7 @@ goog.require('ol.net');
 goog.require('ol.proj');
 goog.require('ol.source.State');
 goog.require('ol.source.TileImage');
+goog.require('ol.tilegrid');
 
 
 /**
