+++ test/spec/ol/source/urltile.test.js
@@ -2,6 +2,7 @@ goog.provide('ol.test.source.UrlTile');
 
 goog.require('ol.proj');
 goog.require('ol.source.UrlTile');
+goog.require('ol.tilegrid');
 
 
 describe('ol.source.UrlTile', function() {
