+++ test/spec/ol/source/xyz.test.js
@@ -4,6 +4,7 @@ goog.require('ol.source.Tile');
 goog.require('ol.source.TileImage');
 goog.require('ol.source.UrlTile');
 goog.require('ol.source.XYZ');
+goog.require('ol.tilegrid');
 
 
 describe('ol.source.XYZ', function() {
