+++ test/spec/ol/tileurlfunction.test.js
@@ -1,6 +1,7 @@
 goog.provide('ol.test.TileUrlFunction');
 
 goog.require('ol.TileUrlFunction');
+goog.require('ol.tilegrid');
 goog.require('ol.tilegrid.TileGrid');
