+++ test_rendering/spec/ol/reproj/image.test.js
@@ -5,6 +5,7 @@ goog.require('ol.proj');
 goog.require('ol.proj.EPSG3857');
 goog.require('ol.reproj.Image');
 goog.require('ol.source.ImageStatic');
+goog.require('ol.tilegrid');
 
 
 describe('ol.rendering.reproj.Image', function() {
