+++ test_rendering/spec/ol/reproj/tile.test.js
@@ -1,10 +1,11 @@
 goog.provide('ol.test.rendering.reproj.Tile');
 
+goog.require('ol.TileState');
 goog.require('ol.events');
 goog.require('ol.proj');
 goog.require('ol.reproj.Tile');
 goog.require('ol.source.XYZ');
-goog.require('ol.TileState');
+goog.require('ol.tilegrid');
 
 
 describe('ol.rendering.reproj.Tile', function() {
